+++ frontend/lost/src/containers/Datasets/Datasets.jsx
@@ -3,7 +3,7 @@ import { faFolderPlus } from '@fortawesome/free-solid-svg-icons'
 import { useEffect, useState } from 'react'
 import { useToggle } from 'react-use'
 import * as annotaskApi from '../../api/anno_task'
-import * as datasetApi from '../../actions/dataset/dataset_api'
+import * as datasetApi from '../../api/dataset_api'
 import DatasetEditModal from './DatasetEditModal'
 import DatasetExportModal from './DatasetExportModal'
 import DatasetTable from './DatasetTable'
