+++ frontend/lost/src/containers/pipeline/start/modals/anno-task-modal/steps/SelectStorageSettings.tsx
@@ -3,7 +3,7 @@ import { faBoxesPacking } from '@fortawesome/free-solid-svg-icons'
 import { useNodesData, useReactFlow } from '@xyflow/react'
 import { useState } from 'react'
 import Select from 'react-select'
-import * as datasetApi from '../../../../../../actions/dataset/dataset_api'
+import * as datasetApi from '../../../../../../api/dataset_api'
 import DatasetEditModal from '../../../../../Datasets/DatasetEditModal'
 import { AnnoTaskNodeData } from '../../../nodes'
 import CoreIconButton from '../../../../../../components/CoreIconButton'
